Test Level Blockout Complete
File in 'Scenes' folder and called 'Train_Blockout'. I've made a test level to walk around in & added some geometry that the player can traverse. The start of the scene in the underneath of a trainstation and after traversing above ground you make it onto the platforms. I was hoping this area to be prime for some sort of shootout etc :)