Text
Made a new prefab called "NoTextSayDialogue" that will remove the text but still allow the script and VA to function.

Also have "NoTextboxSayDialogue" which still has the text but not the textbox - this is for the dialogue. I need to change the dialogue blocks to have this instead of the "NoTextSayDialogue" Prefab.